+++ tests/resources/metrics_dumps/test_write_all_metrics_by_city.py
@@ -63,7 +63,7 @@ def test_write_CanopyAreaPerResidentInformal(target_folder):
 @timeout_decorator.timeout(SLOW_TEST_TIMEOUT_SECONDS)
 @pytest.mark.skipif(DUMP_RUN_LEVEL != DumpRunLevel.RUN_FAST_ONLY, reason=f"Skipping since DUMP_RUN_LEVEL set to {DUMP_RUN_LEVEL}")
 def test_write_Era5MetPreprocessingUmep(target_folder):
-    metric_obj = Era5MetPreprocessingUmep(seasonal_utc_offset=-3)
+    metric_obj = Era5MetPreprocessingUmep(start_date='2023-01-01', end_date='2023-12-31', seasonal_utc_offset=-3)
     _run_write_metrics_by_city_test(metric_obj, target_folder, GEOEXTENT_TERESINA, GEOZONE_TERESINA)
 
 @timeout_decorator.timeout(SLOW_TEST_TIMEOUT_SECONDS)